Original dataloader file

+++ caltech256.py
@@ -59,7 +59,7 @@ class Caltech256(Dataset):
     def __getitem__(self, idx):
         img_name, label = list(self.files.items())[idx]
         image = PIL.Image.open(img_name).convert("RGB") # A few images are grayscale
-        label = torch.Tensor([label]).type(torch.LongTensor)
+        label = torch.Tensor([label])
 
 
         if self.transform:
